requirements.txt
python-dotenv~=1.0.1
google-generativeai
Flask~=3.0.1
Werkzeug~=3.0.1
gunicorn
bcrypt~=4.1.2
boto3~=1.34.55
pandas~=2.2.1
openpyxl~=3.1.2